+++ frontend/src/components/Dash/Settings/Account.vue
@@ -1,6 +1,6 @@
 <template>
-    <div class="settings">
-        <div class="account">
+    <div class="settings-section">
+        <div class="account-information">
             <h2>Account</h2>
             <p>Here you can edit your login information and username.</p>
             <div class="inputs">
@@ -15,40 +15,6 @@
                     <div :disabled="about.data.demo" @click="changeSetting('Username')">
                         {{ account.account.username }}
                     </div>
-                </div>
-            </div>
-        </div>
-        <div class="security">
-            <h2>Security</h2>
-            <p>Here you can change security related settings such as changing your password or enabling 2FA.</p>
-            <div class="inputs">
-                <div class="input">
-                    <label>Password</label>
-                    <div :disabled="about.data.demo" @click="changeSetting('Password')">
-                        ***********
-                    </div>
-                </div>
-                <div class="input totp">
-                    <label>2FA settings</label>
-                    <button :disabled="about.data.demo" @click="toggleTOTP">
-                        {{ account.account.totp ? "Disable" : "Enable" }} 2FA
-                    </button>
-                </div>
-            </div>
-        </div>
-        <div class="sole-user">
-            <h2>Automatic login</h2>
-            <p>
-                All requests will be automatically authenticated by this account.<br>
-                Requires <a href="https://docs.getlynx.dev/installation/environment-variables" target="_blank" rel="noopener noreferrer">SOLE_USER</a>
-                to be set to <code style="background-color: var(--bg-color-2); padding: 0.2em; border-radius: 5px;">{{  account.account.id }}</code><br>
-                Do not enable unless you have alternative authentication methods in place.
-            </p>
-            <div class="inputs" style="margin-top: 0">
-                <div class="input automatic-login">
-                    <button @click="toggleAutomaticLogin">
-                        {{ account.account.allowAutomaticLogin ? "Disable" : "Enable" }}
-                    </button>
                 </div>
             </div>
         </div>
@@ -72,40 +38,14 @@
                 </div>
             </div>
         </div>
-        <div class="about">
-            <h2>About</h2>
-            <p>Here you can see information about your Lynx installation</p>
-            <table class="about-table">
-                <tr>
-                    <td>Version</td>
-                    <td>{{ about.data.version }}</td>
-                </tr>
-                <tr>
-                    <td>Account Role</td>
-                    <td>{{ account.account.role }}</td>
-                </tr>
-                <tr>
-                    <td>Account ID</td>
-                    <td>{{ account.account.id }}</td>
-                </tr>
-                <tr v-if="Object.prototype.hasOwnProperty.call(about.data, 'links')">
-                    <td>Links</td>
-                    <td>{{ formatNumber(about.data.links) }}</td>
-                </tr>
-                <tr v-if="Object.prototype.hasOwnProperty.call(about.data, 'accounts')">
-                    <td>Accounts</td>
-                    <td>{{ formatNumber(about.data.accounts) }}</td>
-                </tr>
-            </table>
-        </div>
     </div>
 </template>
 
 <script>
-import { useAccountStore } from "../../stores/account";
-import { usePopups } from "../../stores/popups";
-import { useAbout } from "../../stores/about";
-import SecretBox from "./SecretBox.vue";
+import { useAccountStore } from "../../../stores/account";
+import { usePopups } from "../../../stores/popups";
+import { useAbout } from "../../../stores/about";
+import SecretBox from "../SecretBox.vue";
 
 export default {
     components: {
@@ -122,45 +62,6 @@ export default {
         changeSetting(name) {
             if (this.about.data.demo) return;
             this.popups.addPopup(`Change${name}`, { account: this.account.account.id });
-        },
-        async toggleTOTP() {
-            if (this.about.data.demo) return;
-            const totpEnabled = this.account.account.totp;
-            if (!totpEnabled) {
-                this.popups.addPopup("EnableTOTP", {});
-            } else {
-                this.popups.addPopup("DisableTOTP", {});
-            }
-        },
-        async toggleAutomaticLogin() {
-            const verificationData = await this.popups.addPopup("Verify", { async: true });
-            const loadingPopup = await this.popups.addPopup("Loader");
-
-            const response = await this.account.fetch("/auth/automatic-login", {
-                method: this.account.account.allowAutomaticLogin ? "DELETE" : "POST",
-                body: JSON.stringify({
-                    verification: verificationData,
-                }),
-            });
-
-            this.popups.closePopup(loadingPopup.id);
-
-            if (!response.success) {
-                this.popups.addPopup("Information", {
-                    title: `Error ${this.account.account.allowAutomaticLogin ? "disabling" : "enabling"} automatic login`,
-                    description: response.message,
-                    buttons: [
-                        {
-                            name: "Okay",
-                            type: "primary",
-                            action: "close-all",
-                        },
-                    ],
-                });
-                return;
-            }
-
-            this.account.getAccount();
         },
         async getConfig() {
             const data = await this.account.fetch("/sharex/config", {
@@ -195,16 +96,12 @@ export default {
 
             document.body.removeChild(a);
         },
-        formatNumber(number) {
-            return number.toLocaleString();
-        },
     },
 };
 </script>
 
 .<style lang="scss" scoped>
-.settings {
-    padding-block: 2rem;
+.settings-section {
     text-align: left;
     display: flex;
     flex-direction: column;
@@ -344,36 +241,9 @@ export default {
                 }
             }
         }
-        &.about {
-            table.about-table {
-                border-collapse:separate;
-                width: max-content;
-                border: 1px solid var(--bg-color-3);
-                border-radius: 10px;
-                tr, th {
-
-                    border-radius: 10px;
-                }
-                tr {
-                    &:nth-of-type(2n) {
-                        background: var(--bg-color-2);
-                    }
-                    td {
-                        padding: 0.6rem 0.8rem;
-                        &:nth-of-type(1) {
-                            font-weight: 500;
-                        }
-                        &:nth-of-type(2) {
-                            font-weight: 300;
-                        }
-                    }
-                }
-            }
-        }
     }
 
     @media screen and (max-width: 768px) {
-        padding-inline: 2rem;
         > div {
             h2 {
                 font-size: 3rem;