+++ frontend/src/main.js
@@ -30,6 +30,7 @@ import {
     faUser,
     faEnvelope,
     faKey,
+    faCircleInfo,
 } from "@fortawesome/free-solid-svg-icons";
 import VueObserveVisibility from "vue3-observe-visibility";
 import ContextMenu from "@imengyu/vue3-context-menu";
@@ -67,6 +68,7 @@ library.add(
     faUser,
     faEnvelope,
     faKey,
+    faCircleInfo,
 );
 
 const app = createApp(App);