+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - password
 - Improve responsiveness for narrow (but non-mobile) screens
 - Consistently use array format for fontawesome icons
+- Add tabbed layout to Settings and split into own components
 - Bug fixes:
   - Admins were not able to update standard user's roles, this has been resolved.
   - Admins are no longer able to see other admin's links.